+++ v3/integrations/nrsnowflake/nrsnowflake.go
@@ -49,8 +49,8 @@ import (
 	"strconv"
 
 	"github.com/newrelic/go-agent/v3/internal"
-	"github.com/newrelic/go-agent/v3/internal/sqlparse"
 	newrelic "github.com/newrelic/go-agent/v3/newrelic"
+	"github.com/newrelic/go-agent/v3/newrelic/sqlparse"
 	"github.com/snowflakedb/gosnowflake"
 )